pyproject.toml
[tool.poetry]
name = "get-papers-list"
version = "0.1.0"
description = "A CLI tool to fetch PubMed research papers with non-academic authors (e.g., pharmaceutical or biotech affiliations)."
authors = ["Vinothraja <vinothraja.t3@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "get_papers" }]
keywords = ["pubmed", "biotech", "pharma", "CLI", "research", "filter"]

[tool.poetry.dependencies]
python = "^3.10"
biopython = "^1.83"
pandas = "^2.2.2"
click = "^8.1.7"
tabulate = "^0.9.0"
ratelimit = "^2.2.1"
rich = "^14.0.0"

[tool.poetry.scripts]
get-papers-list = "get_papers.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"